renaming path to id and moving to correct class
