+++ Admin/StaticContentAdmin.php
@@ -28,7 +28,7 @@ class StaticContentAdmin extends Admin
     protected function configureListFields(ListMapper $listMapper)
     {
         $listMapper
-            ->addIdentifier('path', 'text')
+            ->addIdentifier('id', 'text')
             ->add('title', 'text')
         ;
     }
